feat: add capability to the solution to enable users to deploy either jira or service-now integration with the respective parameters; made updates in the cdk and app, created a bash script; fix tests
