+++ app_jira.py
@@ -1,0 +1,15 @@
+#!/usr/bin/env python3
+import os
+import aws_cdk as cdk
+from aws_security_incident_response_sample_integrations.aws_security_incident_response_sample_integrations_common_stack import AwsSecurityIncidentResponseSampleIntegrationsCommonStack
+from aws_security_incident_response_sample_integrations.aws_security_incident_response_jira_integration_stack import AwsSecurityIncidentResponseJiraIntegrationStack
+
+app = cdk.App()
+
+# Create common stack
+common_stack = AwsSecurityIncidentResponseSampleIntegrationsCommonStack(app, "AwsSecurityIncidentResponseSampleIntegrationsCommonStack")
+
+# Create Jira integration stack
+jira_stack = AwsSecurityIncidentResponseJiraIntegrationStack(app, "AwsSecurityIncidentResponseJiraIntegrationStack", common_stack=common_stack)
+
+app.synth()
